requirement.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.1
django-ckeditor-5==0.2.18
django-filter==25.1
django-jalali-date==2.0.0
django-js-asset==3.1.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-solo==2.4.0
django-stubs-ext==5.2.0
django-summernote==0.8.20.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jalali_core==1.0.0
jdatetime==5.2.0
laces==0.1.2
moviepy==2.2.1
numpy==2.2.6
openpyxl==3.1.5
pillow==11.2.1
pillow_heif==0.21.0
proglog==0.1.12
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==78.1.1
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
wheel==0.45.1
Willow==1.10.0
